First Steps
First steps in the implementation of Conway’s Game of life.
The system still does not behave as it should.

Cell class is very basic and should transform later into an abstract
class. For now it mostly holds a boolean variable and some functions to
operate on it. Cell objects are passive and are not aware of anything
but themselves.

Field class is the heavy lifter. It is the main data structure. It
generates the grid and populate it with “dead” cells. It also holds all
the methods to operate on the cells.

GoL (Game of Life) class is the one holding the main() method and
interacts with the user. Basic console UI.